I had to modify the build order to have the layout extension build before the sbml directory since the objects in the layout directory are needed to build the sbml directory. So basically I added sbml/layout to Makefile.in in src and removed it from Makefile.in in src/sbml.
There also was a typo in one Makefile.in (subdires -> subdirs).

I added three example programs that use the layout extension.
I used them to test if the extesnion actually works since there are
no unit tests for the layout extension yet.
So far the three example files generate the correct output.

I also had to modify the Makefile.in of the examples directory to include
config/makefile-common-vars.mk so that USE_LAYOUT is defined.