Modified a constructor 'UnitFormulaFormatter::UnitFormulaFormatter(const Model *m)'.
Now, the given Model poitner is directly set to 'model' variable (a data member of 
UnitFormulaFormatter). In other words, m->clone() is not called now.
I think we don't have to set a copy of the given Model object (m->clone()) to 'model' 
variable because the 'model' variable is const and therefore never changed during 
the lifetime of an UnitFormulaFormatter object.

This change may drastically reduce consumed memory and calculation time
required for Unit consistency check functions when the given model is large, 
because the overhead of m->clone() could be large and the constructor of 
UnitFormulaFormatter could be heavily called in the functions (especially in the 
functions which depend on recursive calls).