Fixed SBMLErrorLog::remove(const unsigned int) function to remove only the first matched SBMLError object instead of to remove all the matched SBMLError objects.
At first, I though this function removes *all* the matched SBMLError objects 
which have the given errorId, and I reimplemented the function to behave so 
in Revision 8759.
However, I just realized that this is not true, and actually this
function was originally implemented to remove only one SBMLError object even 
if there are multiple SBMLError objects which has the given errorId.

Due to this problem, a segmentation fault could be occurred when reading
some invalid SBML file.