more sorting out constructors and NULL references
Also, it was increasingly annoying me that the assignment operator functions sometimes used &orig as an argument and sometimes used &rhs. In some cases it was even inconsistent between the function source and the function header. They are now all rhs - except ones I havent changed yet but they will be.