Heinous hack #244:
We now define a new command, @sbmlpackage, to use instead of @ingroup.  The command is part of a hack that selectively hides the use of @ingroup in a two-pass scheme of running Doxygen.  In the first pass, @ingroup is redefined as @addindex, chosen only because it's a Doxygen command that takes an argument but otherwise has no effect on the output we generate.  We ultimately need to use @ingroup, but when it's used for grouping classes, Doxygen (at least in version 1.8.5) fails to include the classes in the table of contents for the navigation panel along the left-hand side of the documentation pages.  To work around this problem, we do a first Doxygen pass where we hide the use of @ingroup, using this hack of redefining @sbmlpackage as @addindex (which is a no-op for us).  In the second (final) pass, we expose the @ingroup command by defining @sbmlpackage to be @ingroup.  This results in the table of contents lacking all the classes flagged with @ingroup.  To solve that, we copy over the index files created from the first Doxygen pass.

This is all implemented in Makefile.in.