Implemented new Doxygen macros that wrap some of the previous commands.  There are now three Doxygen commands:
  @sbmlbrief(package}
  @sbmlmarker{package}
  @sbmlpackage{package}

In most cases, such as for class documentation at the top of class .h files, only the first is needed.  It allows us to replace what we were doing before, involving separate commands such as

  @class CompExtension
  @sbmlpackage comp
  @brief @htmlinclude pkg-marker-comp.html
  The core module of the &ldquo;comp&rdquo; package extension.
 
with just two commands:

  @class CompExtension
  @sbmlbrief{comp} The core module of the &ldquo;comp&rdquo; package
  extension.

This has several advantages: (1) it streamlines the documentation command so that there are fewer opportunities to forget a piece or get the syntax wrong; (2) it abstracts out the styling command (the previous @htmlinclude pkg-marker), which is a better design than mixing presentation with semantic markup; (3) it's shorter, therefore easier to read; and (4) the abstraction opens the opportunity for changing or doing more behind the scenes without having to modify all the files every time.

Aside: the change to make @sbmlpackage use {} around arguments is not strictly needed, but the {} arguments *are* needed for the other commands because of how Doxygen aliases work.  So, I changed the previous version of @sbmlpackage, which didn't use an argument, to use {} for arguments the other commands so that all the @sbml* commands used the same syntax.  (Otherwise we'd have one that uses no curly braces and others that do, which would be an opportunity for errors.)