Essentially rewrote this completely, to update it and add support for generating the table of validator error codes located at http://sbml.org/Facilities/Documentation/Error_Categories
The new version works like this:

  writeErrorTable.py [--doc | --enum | --web]  OUTPUTFILE

The first argument must one of --doc, --enum or --web.

  * Given --doc, it will generate a table suitable for use in the libSBML
    API documentation as it appears in the header of SBMLError.h.

  * Given --enum, it will generate the SBMLErrorCode_t enum for SBMLError.h.

  * Given --web, it will generate output used in the table in the web
    page at http://sbml.org/Facilities/Documentation/Error_Categories.

It will write its output to OUTPUTFILE.  Note that the name of the file
will be used as given; no file name extension will be added.  Callers are
assumed to supply the full name desired.