Fix for https://www.pivotaltracker.com/story/show/140698337
I tried to be too clever with the numeric checks on a functionDefinition and they save information about functions already checked BUT rechecked children if a function was called a second time. With piecewise this is done somewhere else and checking children will return a false negative (since al least one must be boolean). This only comes to light when you call the same piecewise functionDefinition more than once.