Fix for Matthias's issue of a conversion factor's units not matching.  The problem was that the submodel could have a different set of packages enabled, which made libsbml refuse to combine the units (of the conversion factor and of the replaced elements).  To fix this, we instead combine the units of the parent element and the inverse of the conversion factor, and compare that to the units of the replaced element.
Also added a check for non-existent conversion factor: usually this will never fire, but it's possible to run the unit checking separately (as happens in the validation unit tests), which will crash libsbml without the check.

Also also, add the IDs of the elements in question for the error message.